feature/BB-14 - paginator adapted
