--- css/common.css
@@ -1,14 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,400;0,700;1,400;1,700&display=swap");
-* {
-  margin: 0;
-  padding: 0;
-}
-
-html {
-  font-size: 16px;
-}
-
-body {
-  font-family: "Roboto", sans-serif;
-  color: #2d2d2d;
-}
