--- css/shablons.css
@@ -1,1 +1,0 @@
-/* No CSS */
