+++ src/components/atoms/Loader/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import * as styles from "./styles.module.scss";
-import loader from "../../../assets/images/loader.gif";
+import loader from "src/assets/images/loader.gif";
 
 function Loader() {
   return (
