+++ src/components/organisms/CardList/styles.module.scss
@@ -1,7 +1,7 @@
 @import "../../../assets/styles/styles.scss";
 
 .cardList {
-  min-height: calc(100vh - 10rem);
+  //min-height: calc(100vh - 10rem);
   .container {
     @include content;
     display: flex;
