+++ src/components/organisms/Header/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Link, useLocation } from "react-router-dom";
 import * as styles from "./styles.module.scss";
 import { ReactComponent as Logo } from "../../../assets/images/logo.svg";
-import burger from "../../../assets/icons/burger.svg";
+import burger from "src/assets/icons/burger.svg";
 import Btn from "src/components/atoms/Btn";
 
 function Header() {
