+++ src/components/organisms/Paginator/styles.module.scss
@@ -7,5 +7,8 @@
     display: flex;
     align-items: center;
     margin-right: 2rem;
+    @include mobileSmall {
+      margin-right: 0.5rem;
+    }
   }
 }
