+++ src/components/organisms/TitleAndToggle/component.js
@@ -1,10 +1,10 @@
 import React from "react";
 import * as styles from "./styles.module.scss";
 import Btn from "src/components/atoms/Btn/index";
-import lines from "../../../assets/icons/lines.svg";
-import tiles from "../../../assets/icons/tiles.svg";
-import lines2 from "../../../assets/icons/lines2.svg";
-import tiles2 from "../../../assets/icons/tiles2.svg";
+import lines from "src/assets/icons/lines.svg";
+import tiles from "src/assets/icons/tiles.svg";
+import lines2 from "src/assets/icons/lines2.svg";
+import tiles2 from "src/assets/icons/tiles2.svg";
 
 function TitleAndToggle({ tile, changeStyles }) {
   return (
