+++ src/components/pages/Main/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import * as styles from "./styles.module.scss";
 import { LazyLoadImage } from "react-lazy-load-image-component";
-import image from "../../../assets/images/breaking-bad.webp";
+import image from "src/assets/images/breaking-bad.webp";
 
 function Main() {
   return (
